+++ packages/plugin-client/src/components/index.ts
@@ -1,3 +1,4 @@
+export { ClassClient } from './ClassClient.tsx'
 export { Client } from './Client.tsx'
 export { Operations } from './Operations.tsx'
 export { Url } from './Url.tsx'