+++ packages/plugin-vue-query/src/generators/infiniteQueryGenerator.tsx
@@ -42,8 +42,10 @@ export const infiniteQueryGenerator = createReactGenerator<PluginVueQuery>({
     }
 
     const hasClientPlugin = !!pluginManager.getPluginByKey([pluginClientName])
+    // Class-based clients are not compatible with query hooks, so we generate inline clients
+    const shouldUseClientPlugin = hasClientPlugin && options.client.clientType !== 'class'
     const client = {
-      name: hasClientPlugin
+      name: shouldUseClientPlugin
         ? getName(operation, {
             type: 'function',
             pluginKey: [pluginClientName],
@@ -124,8 +126,8 @@ export const infiniteQueryGenerator = createReactGenerator<PluginVueQuery>({
         )}
         <File.Import name={['toValue']} path="vue" />
         <File.Import name={['MaybeRefOrGetter']} path="vue" isTypeOnly />
-        {hasClientPlugin && <File.Import name={[client.name]} root={query.file.path} path={client.file.path} />}
-        {!hasClientPlugin && (
+        {shouldUseClientPlugin && <File.Import name={[client.name]} root={query.file.path} path={client.file.path} />}
+        {!shouldUseClientPlugin && (
           <File.Import name={['buildFormData']} root={query.file.path} path={path.resolve(config.root, config.output.path, '.kubb/config.ts')} />
         )}
         <File.Import
@@ -150,7 +152,7 @@ export const infiniteQueryGenerator = createReactGenerator<PluginVueQuery>({
           typeSchemas={type.schemas}
           transformer={options.queryKey}
         />
-        {!hasClientPlugin && (
+        {!shouldUseClientPlugin && (
           <Client
             name={client.name}
             baseURL={options.client.baseURL}