+++ packages/plugin-pinia-colada/CHANGELOG.md
@@ -1,6 +1,6 @@
 # @kubb/plugin-pinia-colada
 
-## 4.9.0
+## 4.10.0
 
 ### Minor Changes